% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx2238 - MDS HW5.3.R
\name{correlogram}
\alias{correlogram}
\title{A correlogram Function
This function allows you to produce the the correlogram for all numeric variables in the data, with year given.}
\usage{
correlogram(x)
}
\arguments{
\item{x, }{input year.}
}
\description{
A correlogram Function
This function allows you to produce the the correlogram for all numeric variables in the data, with year given.
}
\examples{
input 1950, comes out the correlogram of all numerical variables
correlogram()

}
\keyword{correlogram}
