% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun1.R
\name{fun1}
\alias{fun1}
\title{Most points scored in given year}
\usage{
fun1(x, y)
}
\arguments{
\item{x}{dataframe (stats2.csv)}

\item{y}{is a year in the dataframe}
}
\description{
This function shows which player scored most points in the year input
}
\examples{
 fun1(stats2, 2000)
}
\keyword{most}
\keyword{points}
