% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun4.R
\name{fun4}
\alias{fun4}
\title{cor}
\usage{
fun4(x, y)
}
\arguments{
\item{x}{dataframe (stats2.csv)}

\item{y}{is a year in the dataframe}
}
\description{
This function shows a corelation (cor) matrix between all numeric variables in the dataset for the given year
}
\examples{
 fun4(stats2, 2000)
}
\keyword{corelation}
\keyword{matrix}
