% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jordan_pts.R
\name{get_jordan_pts}
\alias{get_jordan_pts}
\title{A Cat Function}
\usage{
get_jordan_pts(df, year = 1986)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
get_jordan_pts()
}
\keyword{cats}
