% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oldest_player.R
\name{get_oldest_player}
\alias{get_oldest_player}
\title{A Cat Function}
\usage{
get_oldest_player(df, year = 1986)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
get_oldest_player()
}
\keyword{cats}
