\name{max_pts}
\alias{max_pts}
\title{fun1}
\description{
%%  ~~ function returns the player name and number of points of the player
%% of the player that scored the most in a given year (year is provided by)
%% the user, within the nba dataset
}
\usage{
max_pts(x)
}

\author{
%%  Kevin Lane, kfl2122
}
\examples{
%% max_pts(2003)

