% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx2238 - MDS HW5.2.R
\name{oldest_of_the_year}
\alias{oldest_of_the_year}
\title{A oldest_of_the_year Function
This function allows you to access the oldest player in the NBA data, with year given.}
\usage{
oldest_of_the_year(x)
}
\arguments{
\item{x, }{input year.}
}
\description{
A oldest_of_the_year Function
This function allows you to access the oldest player in the NBA data, with year given.
}
\examples{
input 1950, comes out the oldest player age at that year
oldest_of_the_year()

}
\keyword{oldest_of_the_year}
