% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx2238 - MDS HW5.1.R
\name{player_age}
\alias{player_age}
\title{A player_age Function
This function allows you to access a data point of interest in the dataset for some player, with year given.}
\usage{
player_age(x)
}
\arguments{
\item{x, }{input year.}
}
\description{
A player_age Function
This function allows you to access a data point of interest in the dataset for some player, with year given.
}
\examples{
input 1950, comes out a list of players age at that year
player_age()

}
\keyword{player_age}
