% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_correlation.R
\name{show_correlation}
\alias{show_correlation}
\title{A Cat Function}
\usage{
show_correlation(df, year = 1986)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
show_correlation()
}
\keyword{cats}
