% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearToCor.R
\name{yearToCor}
\alias{yearToCor}
\title{Year to Correlation Function}
\usage{
yearToCor(year)
}
\arguments{
\item{year}{Year of interest.}
}
\description{
This function allows you to see a correlation plot of NBA stats in any year of choice.
}
\examples{
cat_function()
}
\keyword{NBA}
