% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearToOldest.R
\name{yearToOldest}
\alias{yearToOldest}
\title{Year to Oldest Player Function}
\usage{
yearToOldest(year)
}
\arguments{
\item{year}{Year of interest.}
}
\description{
This function allows you to see the older NBA player in any year of choice.
}
\examples{
cat_function()
}
\keyword{NBA}
