% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearToPlayer.R
\name{yearToPlayer}
\alias{yearToPlayer}
\title{Year to Player Function}
\usage{
yearToPlayer(year)
}
\arguments{
\item{year}{Year of interest.}
}
\description{
This function allows you to see how many points a player scored in any year of choice.
}
\examples{
cat_function()
}
\keyword{NBA}
